refactor: Authentication. - Auth endpoint and service were refactored. - Unnecessary files were deleted.
